Сделала автоматизацию сборки
