+++ index.html
@@ -5,11 +5,6 @@
   <meta name="viewport" content="width=device-width,initial-scale=1">
   <title>HTML Academy: Седона index</title>
   <link rel="stylesheet" href="css/style.css">
-  <script>
-    // Picture element HTML5 shiv
-    document.createElement( "picture" );
-  </script>
-  <script src="js/picturefill.js" async></script>
 </head>
 <body class="closed-menu" id="body">
 <div class="page-wrapper">
@@ -275,6 +270,5 @@ function initMap() {
 }
 </script>
 <script src="js/main.js"></script>
-
 </body>
 </html>
