+++ package.json
@@ -15,8 +15,15 @@
   "devDependencies": {
     "autoprefixer": "6.3.3",
     "browser-sync": "2.11.1",
+    "css-mqpacker": "^4.0.1",
     "editorconfig-tools": "0.1.1",
     "gulp": "3.9.1",
+    "gulp-clean": "^0.3.2",
+    "gulp-combine-mq": "^0.4.0",
+    "gulp-concat": "^2.6.0",
+    "gulp-csso": "^1.1.0",
+    "gulp-imagemin": "^2.4.0",
+    "gulp-jsmin": "^0.1.5",
     "gulp-plumber": "1.1.0",
     "gulp-postcss": "6.1.0",
     "gulp-sass": "2.2.0"
